Declare this package to be obsolete
Since version 22.1, released a mere twelve years ago, Emacs has
built-in support for themes.  The old implementation provided by
this package has not seen a commit in ten years.

For some reason this package is in the top 93rd percentile of
downloads from Melpa.  It is unlikely that so may people still
use this package while being fully aware that Emacs ships with
another implementation, but we cannot be sure, so we don't
remove it from Melpa.

But we should also help all those users who have installed this
package by mistake.  For that reason why now display a warning,
which explains the situation, when it is loaded.